--- src/main/java/com/example/Cat.java
@@ -1,21 +1,0 @@
-package com.example;
-
-import java.util.List;
-
-public class Cat {
-
-    Predator predator;
-
-    public Cat(Feline feline) {
-        this.predator = feline;
-    }
-
-    public String getSound() {
-        return "Мяу";
-    }
-
-    public List<String> getFood() throws Exception {
-        return predator.eatMeat();
-    }
-
-}
