--- src/main/java/com/example/Predator.java
@@ -1,9 +1,0 @@
-package com.example;
-
-import java.util.List;
-
-public interface Predator {
-
-    List<String> eatMeat() throws Exception;
-
-}
