+++ src/main/java/ru/yandex/Cat.java
@@ -1,0 +1,26 @@
+package ru.yandex;
+
+import java.util.List;
+
+public class Cat {
+
+    Predator predator;
+
+    public Cat(Feline feline) {
+        this.predator = feline;
+    }
+    public Cat(){
+
+    }
+
+    public String getSound() {
+        return "Мяу";
+    }
+
+    public List<String> getFood() throws Exception {
+        return predator.eatMeat();
+    }
+
+    public void eatMeat() {
+    }
+}
