+++ src/main/java/ru/yandex/FelineInterface.java
@@ -1,8 +1,9 @@
-package com.example;
+package ru.yandex;
 
 import java.util.List;
 
 public interface FelineInterface {
+    int getKittens();
