+++ src/main/java/ru/yandex/Lion.java
@@ -1,0 +1,37 @@
+package ru.yandex;
+
+import java.util.List;
+
+public class Lion extends Feline{
+    private FelineInterface felineInterface;
+    boolean hasMane;
+
+    public Lion(FelineInterface felineInterface){
+
+        this.felineInterface = felineInterface;
+    }
+
+    public Lion(String sex) throws Exception {
+        if ("Самец".equals(sex)) {
+            hasMane = true;
+        } else if ("Самка".equals(sex)) {
+            hasMane = false;
+        } else {
+            throw new Exception("Используйте допустимые значения пола животного - самец или самка");
+        }
+    }
+
+    Feline feline = new Feline();
+
+    public int getKittens() {
+        return feline.getKittens();
+    }
+
+    public boolean doesHaveMane() {
+        return hasMane;
+    }
+
+    public List<String> getFood() throws Exception {
+        return feline.getFood("Хищник");
+    }
+}
