+++ src/main/java/ru/yandex/Predator.java
@@ -1,0 +1,9 @@
+package ru.yandex;
+
+import java.util.List;
+
+public interface Predator {
+
+    List<String> eatMeat() throws Exception;
+
+}
