+++ src/test/java/ParamTestLion.java
@@ -1,9 +1,8 @@
-import com.example.Lion;
+import ru.yandex.Lion;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
-import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
